chore(open-cv-test): consolidate executables
