--- tools/opencv-test/src/bin/play.rs
@@ -1,38 +1,0 @@
-// verified to work using a known good mp4 file
-
-use clap::Parser;
-use opencv::{highgui, prelude::*, videoio};
-
-#[derive(Parser, Debug)]
-struct Args {
-    /// the video file to play
-    input: String,
-}
-
-#[tokio::main]
-async fn main() -> anyhow::Result<()> {
-    let args = Args::parse();
-
-    let window = "video playback";
-    highgui::named_window(window, highgui::WINDOW_AUTOSIZE)?;
-    let mut cam = videoio::VideoCapture::from_file(&args.input, videoio::CAP_ANY)?;
-    let opened = videoio::VideoCapture::is_opened(&cam)?;
-    if !opened {
-        panic!("Unable to read from file!");
-    }
-
-    loop {
-        let mut frame = Mat::default();
-        cam.read(&mut frame)?;
-        if frame.size()?.width > 0 {
-            highgui::imshow(window, &frame)?;
-        }
-
-        let key = highgui::wait_key(10)?;
-        if key > 0 && key != 255 {
-            break;
-        }
-    }
-
-    Ok(())
-}
